My personal website
I would like to explain the pages found here:
On the Almanca.html page, there are daily German words and their corresponding Turkish words. It changes daily and gives 10 words daily.

In Cekilis.html, two users are asked to enter numbers between 1 and 100 in the inputs on the page, and then the button is pressed. The system generates a random number after the button is pressed, and the closest number entered by the users is considered the winner.

In the finans.html section, enter your monthly income in the specific field and it will show the amount you need to save monthly and the amount you need to allocate for entertainment. This system is a personal example and it is up to you whether you use it or not.
The giris.html and registration.html pages are currently inactive. Because it's my first project, I can't save data, so it doesn't have functions.

oyun.html is a dice game. There are two buttons on the page and one rolls the dice and the other resets the dice values ​​on the screen.